requirements.txt
discord.py
edge-tts
python-dotenv
langdetect
PyNaCl
